+++ crates/cli/src/commands/graph/project.rs
@@ -4,6 +4,7 @@ use moon::{build_project_graph, load_workspace};
 pub async fn project_graph(
     project_id: &Option<String>,
     dot: bool,
+    json: bool,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let mut workspace = load_workspace().await?;
     let mut project_build = build_project_graph(&mut workspace)?;
@@ -22,8 +23,15 @@ pub async fn project_graph(
         return Ok(());
     }
 
-    let (server, mut tera) = setup_server().await?;
     let graph_info = project_graph_repr(&project_graph).await;
+
+    if json {
+        println!("{}", serde_json::to_string(&graph_info)?);
+
+        return Ok(());
+    }
+
+    let (server, mut tera) = setup_server().await?;
     let url = format!("http://{}", server.server_addr());
     let _ = open::that(&url);